Move tests to test_model
